prescription queue: show full location
